[add] mixin: basic-picker + picker
