+++ src/common/mixins/picker.js
@@ -1,22 +1,5 @@
-const DEFAULT_KEYS = {
-  value: 'value',
-  text: 'text'
-}
-
 export default {
   props: {
-    data: {
-      type: Array,
-      default() {
-        return []
-      }
-    },
-    selectedIndex: {
-      type: Array,
-      default() {
-        return []
-      }
-    },
     title: {
       type: String
     },
@@ -32,30 +15,8 @@ export default {
       type: Number,
       default: 2500
     },
-    alias: {
-      type: Object,
-      default() {
-        return {}
-      }
-    },
     zIndex: {
       type: Number
-    }
-  },
-  computed: {
-    valueKey() {
-      return this.alias.value || DEFAULT_KEYS.value
-    },
-    textKey() {
-      return this.alias.text || DEFAULT_KEYS.text
-    }
-  },
-  watch: {
-    data(newVal) {
-      this.setData(newVal, this.selectedIndex)
-    },
-    selectedIndex(newVal) {
-      this.setData(this.data, newVal)
     }
   }
 }
